Fix corridor imports for data constants
