+++ src/main.js
@@ -6,8 +6,6 @@ import {
   merchantSprites,
   DOOR_SPRITES,
   BESTIARY_PAGES,
-  MANOR_KEY_CONSUMABLE_KEY,
-  ENHANCED_DOOR_CHANCE,
   ENHANCED_GOLD_MULTIPLIER,
   RELIC_DEFINITIONS,
   CONSUMABLE_DEFINITIONS,
@@ -16,17 +14,12 @@ import {
   MEMORY_DEFINITIONS,
   MEMORY_MAP,
   ROOM_DEFINITIONS,
-  ROOM_MAP,
-  DOOR_CATEGORIES,
   FOYER_ROOM,
-  TOTAL_ROOMS_PER_RUN,
-  ROOMS_BEFORE_BOSS,
 } from "./data/index.js";
 import {
   adjustConsumableCount,
   adjustGold,
   awardRelic,
-  clearActiveCombat,
   clearCodexView,
   createState,
   ensureCodexSelections,
@@ -69,6 +62,7 @@ import {
   createCoreContribution,
   createMultiCoreContribution,
 } from "./combat/actions.js";
+import screenModules from "./ui/screens/index.js";
 
 (function () {
 
@@ -78,768 +72,10 @@ import {
     merchantDraftCost: MERCHANT_BASE_DRAFT_COST,
   });
 
-  function buildInitialRoomPool() {
-    return ROOM_DEFINITIONS.map((room) => room.key);
-  }
-
-  function initializeRunState() {
-    closeCodexOverlay();
-    updateState({
-      roomPool: buildInitialRoomPool(),
-      roomHistory: [],
-      currentRoomNumber: 0,
-      currentRoomKey: null,
-      corridorRefreshes: 0,
-      currentEncounterType: null,
-      currentEncounter: null,
-      playerMemories: [],
-      playerRelics: [],
-      playerGold: 0,
-      playerConsumables: {},
-      playerEssence: DEFAULT_PLAYER_STATS.maxEssence,
-      playerMaxEssence: DEFAULT_PLAYER_STATS.maxEssence,
-      shroudGuardCharges: 0,
-      draftPacks: [],
-      selectedDrafts: [],
-      codexSelections: {},
-      activeScreenContext: null,
-      merchantDraftCost: MERCHANT_BASE_DRAFT_COST,
-      roomRewardsClaimed: {},
-      currentRoomIsEnhanced: false,
-    });
-    clearCodexView();
-    clearActiveCombat();
-  }
-
-  function clearRunState() {
-    closeCodexOverlay();
-    updateState({
-      roomPool: [],
-      roomHistory: [],
-      currentRoomNumber: 0,
-      currentRoomKey: null,
-      corridorRefreshes: 0,
-      lastRunScreen: null,
-      hasSave: false,
-      inRun: false,
-      currentEncounterType: null,
-      currentEncounter: null,
-      playerMemories: [],
-      playerRelics: [],
-      playerGold: 0,
-      playerConsumables: {},
-      playerEssence: DEFAULT_PLAYER_STATS.maxEssence,
-      playerMaxEssence: DEFAULT_PLAYER_STATS.maxEssence,
-      shroudGuardCharges: 0,
-      draftPacks: [],
-      selectedDrafts: [],
-      codexSelections: {},
-      activeScreenContext: null,
-      merchantDraftCost: MERCHANT_BASE_DRAFT_COST,
-      roomRewardsClaimed: {},
-      currentRoomIsEnhanced: false,
-    });
-    clearCodexView();
-    clearActiveCombat();
-  }
-
-  function getDoorCategoryOptions(count) {
-    const categories = sampleWithoutReplacement(DOOR_CATEGORIES, count);
-    return categories.map((category) => ({ ...category }));
-  }
-
-  async function goToRoom(ctx, roomKey, options = {}) {
-    const room = ROOM_MAP.get(roomKey);
-    if (!room) {
-      ctx.showToast("That path is sealed.");
-      return;
-    }
-
-    const nextRoomPool = ctx.state.roomPool.filter((key) => key !== roomKey);
-    updateState({ roomPool: nextRoomPool });
-    let roomHistory = ctx.state.roomHistory.slice();
-    if (roomHistory[roomHistory.length - 1] !== roomKey) {
-      roomHistory = [...roomHistory, roomKey];
-    }
-    updateState({
-      roomHistory,
-      currentRoomNumber: roomHistory.length,
-      currentRoomKey: roomKey,
-      lastRunScreen: "room",
-      corridorRefreshes: 0,
-    });
-
-    const encounterType = options.encounterType || null;
-    updateState({
-      currentEncounterType: encounterType,
-      currentRoomIsEnhanced: !!options.enhanced,
-    });
-    const encounter = getEncounterForType(encounterType);
-    updateState({ currentEncounter: encounter });
-
-    await ctx.transitionTo("room", {
-      room,
-      background: room.background,
-      ariaLabel: room.ariaLabel,
-      encounterType,
-      encounter,
-    });
-    ctx.showToast(`You enter ${room.name}.`);
-  }
-
-  async function goToFoyer(ctx) {
-    updateState({ roomPool: [] });
-    let roomHistory = ctx.state.roomHistory.slice();
-    if (roomHistory[roomHistory.length - 1] !== FOYER_ROOM.key) {
-      roomHistory = [...roomHistory, FOYER_ROOM.key];
-    }
-    updateState({
-      roomHistory,
-      currentRoomNumber: roomHistory.length,
-      currentRoomKey: FOYER_ROOM.key,
-      lastRunScreen: "foyer",
-      corridorRefreshes: 0,
-    });
-
-    updateState({
-      currentEncounterType: "boss",
-      currentRoomIsEnhanced: false,
-    });
-    const encounter = getEncounterForType("boss");
-    updateState({ currentEncounter: encounter });
-
-    await ctx.transitionTo("foyer", {
-      encounterType: "boss",
-      encounter,
-    });
-    ctx.showToast("The foyer looms ahead.");
-  }
-
-  const screens = {
-    splash: {
-      key: "splash",
-      background: backgrounds.splash,
-      ariaLabel: "Exterior view of Cebarti Manor at night.",
-      render(ctx) {
-        const wrapper = createElement("div", "screen screen--splash");
-        const panel = createElement("div", "panel panel--splash");
-
-        const title = createElement("h1", "screen__title", "Cebarti Manor");
-        const subtitle = createElement(
-          "p",
-          "screen__subtitle",
-          "A haunted roguelike prototype set within Helen Cebarti's cursed manor."
-        );
-        const prompt = createElement(
-          "button",
-          "button button--primary splash__cta",
-          "Click to Continue"
-        );
-
-        prompt.addEventListener("click", () => ctx.transitionTo("mainMenu"));
-        wrapper.addEventListener("click", (event) => {
-          if (event.target === wrapper) {
-            ctx.transitionTo("mainMenu");
-          }
-        });
-
-        panel.append(title, subtitle, prompt);
-        wrapper.append(panel);
-        return wrapper;
-      },
-    },
-    mainMenu: {
-      key: "mainMenu",
-      type: "menu",
-      ariaLabel: "Wallpaper from inside Cebarti Manor, used for menus.",
-      render(ctx) {
-        const wrapper = createElement("div", "screen screen--menu");
-        const panel = createElement("div", "panel panel--menu");
-
-        const title = createElement("h1", "screen__title", "Cebarti Manor");
-        const subtitle = createElement(
-          "p",
-          "screen__subtitle",
-          "Choose how you will haunt the manor."
-        );
-
-        const menu = createElement("div", "menu");
-
-        const continueBtn = createElement(
-          "button",
-          "button menu__button",
-          "Continue Run"
-        );
-        continueBtn.style.display = ctx.state.hasSave ? "block" : "none";
-        continueBtn.addEventListener("click", () => {
-          if (!ctx.state.hasSave || !ctx.state.lastRunScreen) {
-            ctx.showToast("No run to continue yet.");
-            return;
-          }
-          updateState({ inRun: true });
-          ctx.transitionTo(ctx.state.lastRunScreen);
-        });
-
-        const newRunBtn = createElement(
-          "button",
-          "button menu__button",
-          "New Run"
-        );
-        newRunBtn.addEventListener("click", () => {
-          initializeRunState();
-          updateState({ inRun: true, hasSave: false, lastRunScreen: "well" });
-          ctx.transitionTo("well");
-        });
-
-        const bestiaryBtn = createElement(
-          "button",
-          "button menu__button",
-          "Bestiary"
-        );
-        bestiaryBtn.addEventListener("click", () => showBestiary(ctx));
-
-        const settingsBtn = createElement(
-          "button",
-          "button menu__button",
-          "Settings"
-        );
-        settingsBtn.addEventListener("click", () => ctx.transitionTo("settings"));
-
-        const exitBtn = createElement("button", "button menu__button", "Exit");
-        exitBtn.addEventListener("click", () => {
-          ctx.exitGame();
-        });
-
-        menu.append(continueBtn, newRunBtn, bestiaryBtn, settingsBtn, exitBtn);
-        panel.append(title, subtitle, menu);
-        wrapper.append(panel);
-        return wrapper;
-      },
-    },
-    settings: {
-      key: "settings",
-      type: "menu",
-      ariaLabel: "Wallpaper from inside Cebarti Manor, used for menus.",
-      render(ctx) {
-        const wrapper = createElement("div", "screen screen--menu");
-        const panel = createElement("div", "panel panel--menu");
-
-        const title = createElement("h2", "screen__title", "Settings");
-        const subtitle = createElement(
-          "p",
-          "screen__subtitle",
-          "Configure your haunting experience. Options will arrive in future updates."
-        );
-
-        const backButton = createElement(
-          "button",
-          "button",
-          "Back to Menu"
-        );
-        backButton.addEventListener("click", () => ctx.transitionTo("mainMenu"));
-
-        panel.append(title, subtitle, backButton);
-        wrapper.append(panel);
-        return wrapper;
-      },
-    },
-    well: {
-      key: "well",
-      background: backgrounds.well,
-      ariaLabel: "The Styx well inside Cebarti Manor.",
-      checkpoint: true,
-      render(ctx) {
-        const wrapper = createElement("div", "screen screen--well");
-
-        updateState({ currentEncounterType: null, currentEncounter: null });
-
-        const title = createElement("h2", "screen__title", "The Styx Well");
-        const subtitle = createElement(
-          "p",
-          "screen__subtitle",
-          "Draft three memories to define this run's starting action pool."
-        );
-
-        const wellScene = createWellScene();
-        const draftPanel = renderMemoryDraft(ctx);
-
-        const footer = createElement("div", "screen-footer");
-        const continueButton = createElement(
-          "button",
-          "button button--primary",
-          "Continue"
-        );
-        continueButton.disabled =
-          ctx.state.playerMemories.length !== ctx.state.draftPacks.length;
-        continueButton.addEventListener("click", async () => {
-          if (ctx.state.playerMemories.length !== ctx.state.draftPacks.length) {
-            ctx.showToast("Choose one memory from each set to proceed.");
-            return;
-          }
-          updateState({ corridorRefreshes: 0, lastRunScreen: "corridor" });
-          await ctx.transitionTo("corridor");
-          ctx.showToast("You ascend into the corridor.");
-        });
-        footer.appendChild(continueButton);
-
-        wrapper.append(title, subtitle, wellScene, draftPanel, footer);
-        window.requestAnimationFrame(() => updateMemoryDraftSelection(ctx));
-        return wrapper;
-      },
-    },
-    corridor: {
-      key: "corridor",
-      background: backgrounds.corridor,
-      ariaLabel: "A corridor within Cebarti Manor awaiting door choices.",
-      checkpoint: true,
-      render(ctx) {
-        if (ctx.options && ctx.options.fromRoom) {
-          updateState({ currentRoomKey: null });
-        }
-
-        updateState({ currentEncounterType: null, currentEncounter: null });
-
-        const wrapper = createElement("div", "screen screen--corridor");
-        const availableRooms = Array.isArray(ctx.state.roomPool)
-          ? ctx.state.roomPool
-          : [];
-        const roomsRemaining = Math.max(availableRooms.length, 0);
-        const roomsCleared = Math.min(
-          ROOMS_BEFORE_BOSS - roomsRemaining,
-          ROOMS_BEFORE_BOSS
-        );
-
-        const tracker = createRunTracker(
-          ctx,
-          `Rooms Cleared ${roomsCleared}/${ROOMS_BEFORE_BOSS}`
-        );
-
-        const title = createElement("h2", "screen__title", "The Corridor");
-        const refreshCount = ctx.state.corridorRefreshes;
-        let descriptionText;
-
-        if (roomsRemaining === 0) {
-          descriptionText =
-            "Only the foyer remains. Steel yourself before the final confrontation.";
-        } else if (roomsRemaining === 1) {
-          descriptionText =
-            "Only a single chamber stands between you and the foyer. Choose with care.";
-        } else {
-          descriptionText = `${roomsRemaining} chambers remain before the foyer.`;
-        }
-
-        if (roomsRemaining > 0) {
-          if (refreshCount === 0) {
-            descriptionText += " Three doors shimmer ahead.";
-          } else {
-            descriptionText += ` The manor has reshaped itself ${refreshCount} time${
-              refreshCount === 1 ? "" : "s"
-            }.`;
-          }
-        }
-
-        const subtitle = createElement("p", "screen__subtitle", descriptionText);
-
-        updateState({ currentRoomIsEnhanced: false });
-
-        const doorMap = createElement("div", "door-map");
-        if (roomsRemaining === 0) {
-          const { element: foyerDoor, button: foyerButton } = createDoorChoice(
-            "Door to the Foyer",
-            ["door-button--umbra"],
-            {
-              ariaDescription:
-                "Leads directly to the foyer and the final encounter.",
-              displayLabel: "Foyer",
-              detail: "Face Helen Cebarti's final challenge.",
-            }
-          );
-          foyerButton.addEventListener("click", async () => {
-            foyerButton.disabled = true;
-            await goToFoyer(ctx);
-          });
-          doorMap.appendChild(foyerDoor);
-        } else {
-          const doorCount = Math.min(3, roomsRemaining);
-          const roomsForDoors = sampleWithoutReplacement(availableRooms, doorCount);
-          const categories = getDoorCategoryOptions(doorCount);
-          const doors = roomsForDoors.map((roomKey, index) => {
-            const category = categories[index] || {};
-            const enhanced =
-              doorCount <= 1 ? false : Math.random() < ENHANCED_DOOR_CHANCE;
-            return { roomKey, category, enhanced };
-          });
-
-          if (doors.length > 0 && doors.every((door) => door.enhanced)) {
-            const randomIndex = Math.floor(Math.random() * doors.length);
-            doors[randomIndex].enhanced = false;
-          }
-
-          doors.forEach(({ roomKey, category, enhanced }, index) => {
-            const extraClasses = [];
-            if (typeof category.colorClass === "string" && category.colorClass) {
-              extraClasses.push(category.colorClass);
-            }
-            if (enhanced) {
-              extraClasses.push("door-button--enhanced");
-            }
-
-            const detailParts = [];
-            if (category.detail) {
-              detailParts.push(category.detail);
-            }
-            if (enhanced) {
-              detailParts.push("Locked — requires a Manor Key.");
-            }
-
-            const doorLabel = category.label || `Door ${index + 1}`;
-            const detailText = detailParts.join(" ").trim();
-            const ariaDescriptionParts = [
-              category.ariaDescription || "Leads to an unknown chamber.",
-            ];
-            if (enhanced) {
-              ariaDescriptionParts.push("Locked — requires a Manor Key.");
-            }
-
-            const { element: doorChoice, button: doorButton, iconElement, labelElement, lockElement } =
-              createDoorChoice(doorLabel, extraClasses, {
-                ariaDescription: ariaDescriptionParts.join(" "),
-                detail: detailText.length > 0 ? detailText : undefined,
-                displayLabel: doorLabel,
-                dataset:
-                  typeof category.key === "string"
-                    ? { category: category.key }
-                    : undefined,
-                iconSrc: category.icon,
-              });
-
-            if (iconElement) {
-              iconElement.title = category.label || doorLabel;
-            }
-
-            if (labelElement && enhanced) {
-              labelElement.classList.add("door-option__label--locked");
-              const lockBadge = createElement(
-                "span",
-                "door-option__badge",
-                "Locked"
-              );
-              labelElement.appendChild(lockBadge);
-            }
-
-            if (lockElement) {
-              lockElement.loading = "lazy";
-            }
-
-            let isLocked = enhanced;
-            if (isLocked) {
-              doorButton.classList.add("door-button--locked");
-            }
-            doorButton.dataset.locked = isLocked ? "true" : "false";
-            doorButton.dataset.enhanced = enhanced ? "true" : "false";
-            if (category.key) {
-              doorButton.dataset.roomType = category.key;
-            }
-
-            doorButton.addEventListener("click", async () => {
-              if (isLocked) {
-                const hasKey =
-                  getConsumableCount(MANOR_KEY_CONSUMABLE_KEY) > 0;
-                if (!hasKey) {
-                  ctx.showToast(
-                    "The lock refuses to budge. You'll need a Manor Key."
-                  );
-                  doorButton.classList.add("door-button--shake");
-                  window.setTimeout(() => {
-                    doorButton.classList.remove("door-button--shake");
-                  }, 360);
-                  return;
-                }
-                const spent = spendConsumableCharge(
-                  MANOR_KEY_CONSUMABLE_KEY,
-                  ctx
-                );
-                if (!spent) {
-                  ctx.showToast(
-                    "The key slips from your grasp before the lock yields."
-                  );
-                  return;
-                }
-                ctx.showToast("The Manor Key clicks open the lock.");
-                isLocked = false;
-                doorButton.dataset.locked = "false";
-                doorButton.classList.remove("door-button--locked");
-              }
-              doorButton.disabled = true;
-              await goToRoom(ctx, roomKey, {
-                encounterType: category.key,
-                enhanced,
-              });
-            });
-            doorMap.appendChild(doorChoice);
-          });
-        }
-
-        const footer = createElement("div", "screen-footer");
-        if (roomsRemaining > 0) {
-          const continueButton = createElement(
-            "button",
-            "button button--primary",
-            "Continue Down the Corridor"
-          );
-          continueButton.addEventListener("click", async () => {
-            const corridorRefreshes = (ctx.state.corridorRefreshes || 0) + 1;
-            updateState({
-              corridorRefreshes,
-              lastRunScreen: "corridor",
-            });
-            await ctx.transitionTo("corridor", { refresh: true });
-            ctx.showToast("The corridor rearranges itself.");
-          });
-          footer.appendChild(continueButton);
-        }
-
-        const returnButton = createElement(
-          "button",
-          "button",
-          "Return to Main Menu"
-        );
-        returnButton.addEventListener("click", async () => {
-          const confirmReturn = window.confirm(
-            "Abandon this run and return to the main menu?"
-          );
-          if (!confirmReturn) {
-            return;
-          }
-          clearRunState();
-          await ctx.transitionTo("mainMenu");
-          ctx.showToast(
-            "You withdraw from the corridor and return to the manor's entry hall."
-          );
-        });
-        footer.appendChild(returnButton);
-
-        wrapper.append(tracker, title, subtitle, doorMap, footer);
-        return wrapper;
-      },
-    },
-    room: {
-      key: "room",
-      render(ctx) {
-        const roomData = ctx.options?.room;
-        const encounterType =
-          ctx.options?.encounterType || ctx.state.currentEncounterType;
-        const encounter = ctx.options?.encounter || ctx.state.currentEncounter;
-        const wrapper = createElement("div", "screen screen--room");
-        const roomNumber = Math.max(ctx.state.currentRoomNumber, 1);
-        const tracker = createRunTracker(
-          ctx,
-          `Room ${Math.min(roomNumber, TOTAL_ROOMS_PER_RUN)} of ${TOTAL_ROOMS_PER_RUN}`
-        );
-
-        if (!roomData) {
-          const title = createElement("h2", "screen__title", "The Manor Resists");
-          const subtitle = createElement(
-            "p",
-            "screen__subtitle",
-            "The chosen doorway seals shut. You retreat to the corridor."
-          );
-          const footer = createElement("div", "screen-footer");
-          const backButton = createElement(
-            "button",
-            "button button--primary",
-            "Return to the Corridor"
-          );
-          backButton.addEventListener("click", async () => {
-            updateState({ currentRoomKey: null, lastRunScreen: "corridor" });
-            await ctx.transitionTo("corridor", { refresh: true });
-          });
-          footer.appendChild(backButton);
-          updateState({ currentEncounterType: null, currentEncounter: null });
-          wrapper.append(tracker, title, subtitle, footer);
-          return wrapper;
-        }
-
-        const title = createElement("h2", "screen__title", roomData.name);
-        const subtitle = createElement(
-          "p",
-          "screen__subtitle",
-          roomData.description
-        );
-
-        wrapper.append(tracker, title, subtitle);
-
-        if (isCombatEncounter(encounterType)) {
-          const combatExperience = createCombatExperience(ctx, {
-            room: roomData,
-            encounterType,
-            encounter,
-          });
-          wrapper.append(combatExperience.container, combatExperience.footer);
-          return wrapper;
-        }
-
-        const encounterScene = createEncounterScene({ encounter });
-        const prompt = createElement(
-          "p",
-          "screen__subtitle",
-          getEncounterPrompt(encounterType, encounter)
-        );
-        const footer = createElement("div", "screen-footer");
-        const continueButton = createElement(
-          "button",
-          "button button--primary",
-          "Return to the Corridor"
-        );
-        continueButton.addEventListener("click", async () => {
-          updateState({
-            currentRoomKey: null,
-            lastRunScreen: "corridor",
-            corridorRefreshes: 0,
-            currentEncounterType: null,
-            currentEncounter: null,
-          });
-          await ctx.transitionTo("corridor", { fromRoom: true });
-          ctx.showToast("You slip back into the corridor.");
-        });
-        footer.appendChild(continueButton);
-
-        wrapper.append(encounterScene.scene, prompt);
-
-        if (encounterType === "recovery") {
-          const recoveryResult = applyRecoveryRoomBenefits(ctx, roomData.key);
-          const detailText = recoveryResult
-            ? `You absorb the chamber's lingering calm. Maximum Essence increases by ${recoveryResult.essenceIncrease} and your essence is fully restored.`
-            : "The chamber's restorative energies have already been spent.";
-          wrapper.appendChild(
-            createElement("p", "combat-rewards__detail", detailText)
-          );
-          if (recoveryResult) {
-            ctx.showToast(
-              `Your essence deepens by ${recoveryResult.essenceIncrease}.`
-            );
-          } else {
-            ctx.showToast("You feel as renewed as this chamber allows.");
-          }
-          wrapper.appendChild(footer);
-          return wrapper;
-        }
-
-        if (encounterType === "treasure") {
-          const { panel: rewardsPanel } = createRewardsPanel(ctx, {
-            encounterType: "treasure",
-            continueButton,
-          });
-          if (rewardsPanel) {
-            wrapper.appendChild(rewardsPanel);
-          }
-          wrapper.appendChild(footer);
-          return wrapper;
-        }
-
-        if (encounterType === "merchant") {
-          const merchantPanel = createMerchantPanel(ctx, continueButton);
-          wrapper.append(merchantPanel, footer);
-          return wrapper;
-        }
-
-        wrapper.appendChild(footer);
-        return wrapper;
-      },
-    },
-    foyer: {
-      key: "foyer",
-      background: backgrounds.foyer,
-      ariaLabel: FOYER_ROOM.ariaLabel,
-      checkpoint: true,
-      render(ctx) {
-        const wrapper = createElement("div", "screen screen--room screen--foyer");
-        const roomNumber = Math.max(ctx.state.currentRoomNumber, TOTAL_ROOMS_PER_RUN);
-        const tracker = createRunTracker(
-          `Room ${Math.min(roomNumber, TOTAL_ROOMS_PER_RUN)} of ${TOTAL_ROOMS_PER_RUN}`
-        );
-
-        const title = createElement("h2", "screen__title", FOYER_ROOM.name);
-        const subtitle = createElement("p", "screen__subtitle", FOYER_ROOM.description);
-        const encounter = ctx.options?.encounter || ctx.state.currentEncounter;
-        const encounterScene = createEncounterScene({ encounter });
-        const prompt = createElement(
-          "p",
-          "screen__subtitle",
-          getEncounterPrompt("boss", encounter)
-        );
-
-        const footer = createElement("div", "screen-footer");
-        const continueButton = createElement(
-          "button",
-          "button button--primary",
-          "Challenge the Foyer"
-        );
-        continueButton.addEventListener("click", async () => {
-          continueButton.disabled = true;
-          updateState({
-            hasSave: false,
-            lastRunScreen: null,
-            inRun: false,
-            currentEncounterType: null,
-            currentEncounter: null,
-          });
-          await ctx.transitionTo("victory");
-        });
-        footer.appendChild(continueButton);
-
-        wrapper.append(
-          tracker,
-          title,
-          subtitle,
-          encounterScene.scene,
-          prompt,
-          footer
-        );
-        return wrapper;
-      },
-    },
-    victory: {
-      key: "victory",
-      background: backgrounds.victory,
-      ariaLabel: "Sunlight pours through an open archway as the manor releases you.",
-      render(ctx) {
-        const wrapper = createElement("div", "screen screen--victory");
-
-        const title = createElement("h2", "screen__title", "You Escape the Well");
-        const subtitle = createElement(
-          "p",
-          "screen__subtitle",
-          "Congratulations! You have escaped the well and step once more into the realm of the living."
-        );
-        const clearedRooms = Math.min(
-          ctx.state.roomHistory.filter((key) => key !== FOYER_ROOM.key).length,
-          ROOMS_BEFORE_BOSS
-        );
-        const summary = createElement(
-          "p",
-          "screen__subtitle",
-          `You endured ${clearedRooms} haunted chamber${clearedRooms === 1 ? "" : "s"} before confronting the foyer.`
-        );
-
-        const footer = createElement("div", "screen-footer");
-        const menuButton = createElement(
-          "button",
-          "button button--primary",
-          "Return to Main Menu"
-        );
-        menuButton.addEventListener("click", async () => {
-          clearRunState();
-          await ctx.transitionTo("mainMenu");
-          ctx.showToast("You breathe freely in the manor's entry hall.");
-        });
-        footer.appendChild(menuButton);
-
-        wrapper.append(title, subtitle, summary, footer);
-        return wrapper;
-      },
-    },
-  };
+  const screens = {};
+  screenModules.forEach((screen) => {
+    screens[screen.key] = screen;
+  });
 
   function createDoorButton(label, extraClasses = [], options = {}) {
     const button = document.createElement("button");
@@ -937,6 +173,21 @@ import {
 
     return { element: wrapper, button, labelElement, iconElement, lockElement };
   }
+
+  const screenHelpers = {
+    createRunTracker,
+    createDoorChoice,
+    getConsumableCount,
+    spendConsumableCharge,
+    createEncounterScene,
+    getEncounterPrompt,
+    updateMemoryDraftSelection,
+    createWellScene,
+    renderMemoryDraft,
+    showBestiary,
+    closeCodexOverlay,
+    isCombatEncounter,
+  };
 
   function resetResourceDisplayRegistry() {
     updateState({
@@ -2921,52 +2172,6 @@ import {
     }
   }
 
-  function getRandomItem(source) {
-    if (!Array.isArray(source) || source.length === 0) {
-      return null;
-    }
-    const index = Math.floor(Math.random() * source.length);
-    return source[index];
-  }
-
-  function getEncounterPoolForType(type) {
-    switch (type) {
-      case "combat":
-      case "elite":
-        return enemySprites;
-      case "boss":
-        return bossSprites;
-      case "merchant":
-        return merchantSprites;
-      default:
-        return null;
-    }
-  }
-
-  function getEncounterForType(type) {
-    const pool = getEncounterPoolForType(type);
-    if (!pool) {
-      return null;
-    }
-    const sprite = getRandomItem(pool);
-    if (!sprite) {
-      return null;
-    }
-    const animatedTypes = new Set(["combat", "elite", "boss"]);
-    return {
-      sprite,
-      type,
-      kind:
-        type === "merchant"
-          ? "merchant"
-          : type === "boss"
-          ? "boss"
-          : "enemy",
-      animate: animatedTypes.has(type),
-      enterDelay: 2000,
-    };
-  }
-
   function shouldReduceMotion() {
     if (typeof window === "undefined" || !window.matchMedia) {
       return false;
@@ -3199,6 +2404,7 @@ import {
       options,
       updateResources: updateResourceDisplays,
       exitGame,
+      helpers: screenHelpers,
     };
 
     updateState({ activeScreenContext: context });
@@ -3237,6 +2443,7 @@ import {
       transitionTo,
       showToast,
       options: {},
+      helpers: screenHelpers,
     });
     appendContent(splashContent);
     updateState({ currentScreen: "splash" });
