+++ src/ui/screens/index.js
@@ -1,0 +1,19 @@
+import corridor from "./corridor.js";
+import foyer from "./foyer.js";
+import mainMenu from "./main-menu.js";
+import room from "./room.js";
+import settings from "./settings.js";
+import splash from "./splash.js";
+import victory from "./victory.js";
+import well from "./well.js";
+
+export default [
+  splash,
+  mainMenu,
+  settings,
+  well,
+  corridor,
+  room,
+  foyer,
+  victory,
+];
